pyproject.toml
[tool.poetry]
name = "weather_forecast_in_yaroslavl_bot"
version = "0.1.0"
description = ""
authors = ["SaPer663 <saper663ru@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.10.0"
certifi = "^2021.10.8"
chardet = "^4.0.0"
cheroot = "^8.5.2"
CherryPy = "^18.6.1"
idna = "^3.3"
importlib-resources = "^5.4.0"
"jaraco.classes" = "^3.2.1"
"jaraco.collections" = "^3.4.0"
"jaraco.functools" = "^3.4.0"
"jaraco.text" = "^3.6.0"
lxml = "^4.6.4"
more-itertools = "^8.11.0"
portend = "^3.0.0"
pyTelegramBotAPI = "^4.2.0"
pytz = "^2021.3"
requests = "^2.26.0"
six = "^1.16.0"
soupsieve = "^2.3"
tempora = "^4.1.2"
urllib3 = "^1.26.7"
"zc.lockfile" = "^2.0"
zipp = "^3.6.0"
types-requests = "^2.26.0"
types-beautifulsoup4 = "^4.10.5"
python-decouple = "^3.5"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
pep8-naming = "^0.12.1"
flake8-broken-line = "^0.4.0"
flake8-return = "^1.1.3"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
